Update app.json to use essential-o instead of mini
To solve the issue when deploying to Heroku.

Heroku's error. 
Configure environment
There was an issue setting up your app environment.

We tried to create heroku-postgresql:mini, but received an error from the add-on provider. Try the request again, or log a support ticket and include this message: The mini plan has reached end-of-life. You can try again with a valid plan. See the available plans with: heroku addons:plans heroku-postgresql.